+++ packages/api-client/src/libs/send-request/create-request-operation.ts
@@ -184,9 +184,10 @@ export const createRequestOperation = ({
          * Checks if the response is streaming
          * Unfortunately we cannot check the transfer-encoding header as it is not set by the browser so not quite sure how to test when
          * content-type === 'text/plain' and transfer-encoding === 'chunked'
+         *
+         * Currently we are only checking for server sent events. In OpenApi 3.2.0 streams will be added to the spec
          */
-        const isStreaming = response.headers.get('content-type')?.includes('stream')
-
+        const isStreaming = response.headers.get('content-type')?.startsWith('text/event-stream')
         status?.emit('stop')
 
         const responseHeaders = normalizeHeaders(response.headers, shouldUseProxy(proxyUrl, url))